+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/IntelliJPlatformDependenciesExtension.kt
@@ -1459,6 +1459,28 @@ abstract class IntelliJPlatformDependenciesExtension @Inject constructor(
     )
 
     /**
+     * Adds dependencies on the `test-framework` libraries or their variants, required for testing plugins.
+     *
+     * The versions are determined by the IntelliJ Platform build number.
+     *
+     * @param types Test Framework variant types.
+     * @see TestFrameworkType
+     */
+    fun testFrameworks(vararg types: TestFrameworkType) =
+        testFrameworks(types.asList())
+
+    /**
+     * Adds dependencies on the `test-framework` libraries or their variants, required for testing plugins.
+     *
+     * The versions are determined by the IntelliJ Platform build number.
+     *
+     * @param types Test Framework variant types.
+     * @see TestFrameworkType
+     */
+    fun testFrameworks(types: List<TestFrameworkType>) =
+        types.forEach { testFramework(it) }
+
+    /**
      * Adds a dependency on the `test-framework` library or its variant, required for testing plugins.
      *
      * The version, if absent, is determined by the IntelliJ Platform build number.